made it run
- moved all paths to index
- changed from app.path to bp.path? still don't know why that works
- changed some import statments
